Update keywordize function to parse Array fields
If the client specifies a field that is an Array, the function
is unable to break apart the contents of the array and add them
to the keywords list.

This update checks for the field type and adds its contents to
the keyword set whether it is an Array or a string value.